Reject visitor if number of visits are exceeded.
A new visit happens whenever a visitor goes to a link with the visitor querystring parameter.
When a new visit happens, we increment the number of visits in the Visitor model
and check if the maximum number of visits has been exceeded. If this maximum is exceeded,
the visitor will not be set in the request and the link will not be accessible to the user
(as long as the visitor uuid is not set in session).